pyproject.toml
[project]
name = "nonebot-plugin-bilichat-mygo"

version = "6.0.0"

description = "多种B站链接解析，视频词云，AI总结，你想要的都在 bilichat"
authors = [
    { name = "djkcyl", email = "cyl@cyllive.cn" },
    { name = "Well404", email = "well_404@outlook.com" },
    { name = "Merlin", email = "merlin@merlinblog.cn" },
]
dependencies = [
    "bilireq@git+https://github.com/SK-415/bilireq.git",
    "qrcode>=7.4.2",
    "pillow>=9.5.0",
    "lxml>=4.9.2",
    "nonebot-plugin-localstore>=0.4.1",
    "httpx>=0.28.1",
    "dynrender-skia-opt>=0.3.8",
    "nonebot-plugin-apscheduler>=0.3.0",
    "python-multipart>=0.0.6",
    "nonebot-plugin-alconna>=0.42.3",
    "nonebot-plugin-auto-bot-selector>=0.2.0",
    "nonebot-plugin-waiter>=0.4.0",
    "pypng>=0.20220715.0",
    "nonebot2[fastapi,websockets]>=2.4.2",
    "nonebot-adapter-onebot>=2.4.6",
    "nonebot-plugin-htmlrender>=0.6.6",
]
requires-python = ">=3.12,<4.0"
readme = "README.md"
license = { text = "AGPL3.0" }
[project.optional-dependencies]
extra = [
    "numpy>=1.25.1",
    "nonebot-plugin-sentry>=0.2.2",
    "nonebot-plugin-htmlrender>=0.2.0.3",
    "nonebot-plugin-mongodb>=0.1.0",
]
wordcloud = ["jieba>=0.42.1", "wordcloud>=1.8.2.2"]
summary = ["tiktoken>=0.6.0"]
all = [
    "numpy>=1.25.1",
    "jieba>=0.42.1",
    "wordcloud>=1.8.2.2",
    "nonebot-plugin-htmlrender>=0.2.0.3",
    "nonebot-plugin-mongodb>=0.1.0",
    "nonebot-plugin-sentry>=0.4.1",
    "tiktoken>=0.6.0",
]

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.pdm.dev-dependencies]
dev = [
    "nonebot-adapter-onebot>=2.2.4",
    "nonebot-adapter-mirai2>=0.0.22",
    "nonebot-adapter-qq>=1.0.1",
    "nonebot-adapter-satori>=0.10.3",
]

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"
